Fix build errors: consolidate domain models, fix parsers, DAO methods, UI imports, and theme conflicts
- Remove duplicate DomainModels.kt, consolidate models in individual files
- Fix BankParser interface: remove duplicate BankEntity enum, add default canParse()
- Update all parsers to extend BaseBankParser for shared functionality
- Fix BaseBankParser parseDate() with proper pattern matching
- Add missing DAO methods: count(), getByType(), getByCategory(), deleteOlderThan()
- Fix MovementRepository to use correct DAO method names
- Move ClassificationResult/ClassificationSource to separate file
- Fix EnrichmentPipeline: use domain models, fix toDomain extensions with sortOrder
- Fix UI: add missing Compose imports (fillMaxSize, weight, background, ChevronRight)
- Fix theme: remove conflicting ColorScheme declarations from Color.kt
- Fix NotificationCaptureService imports (ParseResult from domain.model)
- Fix Repositories.kt type mismatches (use AgendaEntryEntity)
- Add sortOrder to CategoryEntity and Category domain model

+++ app/src/main/java/com/finanzas/automatica/data/local/entity/AgendaEntryEntity.kt
@@ -26,7 +26,7 @@ data class AgendaEntryEntity(
     val accountIdentifier: String,
     val displayName: String,
     val defaultCategoryId: Long? = null,
-    val color: Int = 0xFF8D6E63,
+    val color: Int = 0xFF8D6E63.toInt(),
     val origin: String = "MANUAL",
     val createdAt: Long = Instant.now().toEpochMilli(),
     val updatedAt: Long = Instant.now().toEpochMilli()
